test(crypto): extract into file helpers getMasterKey
